requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
cloudinary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
apscheduler
python-multipart
requests
pillow
